Expand user home path from find_file(), call it always regardless of any prefix check
