+++ src/actions.cpp
@@ -47,76 +47,6 @@ static const char* strIntCmd(CAction::T_IntCmdType type)
     return "invalid";
 }
 
-int expand_user_path(const char *path, char *expanded_home_path /*The buffer*/)
-{
-    memset(expanded_home_path, '\0', MAX_PATH); // Assume this much memory is always allocated, hopefully
-    char *home_dir = nullptr;
-
-    if (path[1] == '\0' || path[1] == '/')
-    { // '~/path' case
-        home_dir = getenv("HOME");
-        if (home_dir == nullptr)
-        {
-            home_dir = getenv("USERPROFILE");
-        }
-        if (home_dir != nullptr)
-        {
-            snprintf(expanded_home_path, MAX_PATH - 1, "%s%s", home_dir, path + 1);
-        }
-    }
-    else
-    {
-        const char *first_slash = nullptr;
-        first_slash = strchr(path, '/');  // substring starting from '/'
-        size_t linux_username_limit = 32; // As of now
-        char *username = nullptr;
-        if ((first_slash != NULL) && ((first_slash - (path + 1)) <= linux_username_limit))
-        { // '~/someuser/blah' case
-            username = strndup(path + 1, first_slash - (path + 1));
-        }
-        else
-        { // '~someuser' case
-            username = strndup(path + 1, strlen(path + 1));
-        }
-
-        struct passwd pwd;
-        struct passwd *result;
-        size_t bufsize = sysconf(_SC_GETPW_R_SIZE_MAX);
-        char *buffer = new char[bufsize + 1];
-        int retcode = getpwnam_r(username, &pwd, buffer, bufsize, &result);
-        if (result == NULL)
-        {
-            if (retcode != 0)
-            {
-                errno = retcode;
-            }
-            WARNING_NO("Unable to resolve home path for [%s]\n", path);
-            free(username);
-            return -1;
-        }
-        else
-        {
-            home_dir = result->pw_dir;
-        }
-        delete[] buffer;
-        free(username);
-
-        if (home_dir != nullptr)
-        {
-            if (first_slash != nullptr)
-            { // '~username/path' case
-                snprintf(expanded_home_path, MAX_PATH - 1, "%s%s", home_dir, first_slash);
-            }
-            else
-            { // '~username' case?
-                snprintf(expanded_home_path, MAX_PATH - 1, "%s", home_dir);
-            }
-        }
-    }
-
-    return 1;
-}
-
 const char * CAction::comparatorToString(T_Comparator comp)
 {
     switch(comp) {
@@ -603,19 +533,6 @@ void CAction::setPcapArgs (pcap_pkts  *  P_value)
 
 void CAction::setPcapArgs(const char* P_value)
 {
-    if (P_value[0] == '~')
-    {
-        char pcap_absolute_path[MAX_PATH] = {'\0'};
-        if (expand_user_path(P_value, pcap_absolute_path))
-        {
-            P_value = strndup(pcap_absolute_path, strlen(pcap_absolute_path));
-        }
-        else
-        {
-            ERROR("Couldn't resolve pcap file absolute path: %s", P_value);
-        }
-    }
-
     if (M_pcapArgs != NULL) {
         free(M_pcapArgs);
         M_pcapArgs = NULL;
